uniswap_backend/requirements.txt
requests~=2.26.0
python-dotenv~=0.19.2
djangorestframework==3.13.0
djangorestframework_simplejwt==5.0.0
Django~=4.0
celery==5.2.1
web3~=5.25.0
django-cors-headers==3.11.0
psycopg2-binary==2.9.3
redis==4.0.2
uniswap-python==0.5.4
